+++ CHANGELOG.md
@@ -15,7 +15,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 - `parameter_dir` is now configurable to match the existing `template_dir`.
 - A `parameter_files` array allows configuring an array of *explicit* parameter
   files relative to `parameter_dir`. If this option is specified, automatic
-  parameter files based on region and the stack name will be be used.
+  parameter files based on region and the stack name will not be used.
 - A `parameters` hash key allows defining parameters directly on the stack
   definition rather than requiring an external parameter file.